refactor: modularize functions
